+++ bin/test.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# cd quanteverest_analytics
+python3.8 -m unittest discover -s tests/unit/* -p "*_test.py"